Changes needed to move Formik up a level so list and form can use formik values from context.
